split between profileApp and ccipApp
