+++ packages/offchain-resolver/src/http/ccipGateway.ts
@@ -1,50 +1,11 @@
 import * as Lib from 'dm3-lib/dist.backend';
+import ethers from 'ethers';
 import { Signer } from 'ethers';
 import express from 'express';
 import { WithLocals } from './types';
 
 export function ccipGateway(signer: Signer, resolverAddr: string) {
     const router = express.Router();
-
-    router.post(
-        '/',
-        async (req: express.Request & { app: WithLocals }, res, next) => {
-            const { signedUserProfile, name, address } = req.body;
-            const isSchemaValid = Lib.validateSchema(
-                Lib.account.schema.SignedUserProfile,
-                signedUserProfile,
-            );
-
-            //Check if schema is valid
-            if (!isSchemaValid) {
-                return res.status(400).send({ error: 'invalid schema' });
-            }
-
-            const profileIsValid = Lib.account.checkUserProfile(
-                signedUserProfile,
-                address,
-            );
-
-            //Check if profile sig is correcet
-            if (!profileIsValid) {
-                return res.status(400).send({ error: 'invalid profile' });
-            }
-
-            const profileExists = await req.app.locals.db.getUserProfile(name);
-
-            if (profileExists) {
-                return res
-                    .status(400)
-                    .send({ error: 'subdomain already claimed' });
-            }
-            await req.app.locals.db.setUserProfile(
-                name,
-                signedUserProfile.profile,
-            );
-
-            return res.send(200);
-        },
-    );
 
     router.get(
         '/:resolverAddr/:calldata',
