+++ packages/offchain-resolver/src/index.ts
@@ -5,6 +5,9 @@ import cors from 'cors';
 import winston from 'winston';
 import { getDatabase } from './persistance/getDatabase';
 import { ccipGateway } from './http/ccipGateway';
+import { getWeb3Provider } from './utils/getWeb3Provider';
+import { getSigner } from './utils/getSigner';
+import { readKeyFromEnv } from './utils/readKeyEnv';
 
 const app = express();
 app.use(express.json({ limit: '50mb' }));
@@ -21,6 +24,10 @@ app.use(bodyParser.json());
     });
 
     app.locals.db = await getDatabase();
+    app.locals.web3Provider = getWeb3Provider();
 
-    app.use(ccipGateway());
+    const signer = getSigner();
+    const resolverAddress = readKeyFromEnv('RESOLVER_ADDR');
+
+    app.use(ccipGateway(signer, resolverAddress));
 })();
