+++ packages/offchain-resolver/src/utils/getSigner.ts
@@ -1,0 +1,6 @@
+import ethers from 'ethers';
+import { readKeyFromEnv } from './readKeyEnv';
+export function getSigner() {
+    const privateKey = readKeyFromEnv('SIGNER_PRIVATE_KEY');
+    return new ethers.Wallet(privateKey);
+}
