+++ packages/offchain-resolver/src/utils/getWeb3Provider.ts
@@ -1,0 +1,7 @@
+import ethers from 'ethers';
+import { readKeyFromEnv } from './readKeyEnv';
+
+export function getWeb3Provider(): ethers.providers.BaseProvider {
+    const rpc = readKeyFromEnv('RPC');
+    return new ethers.providers.JsonRpcProvider(rpc);
+}
