+++ packages/offchain-resolver/src/utils/readKeyEnv.ts
@@ -1,0 +1,8 @@
+export function readKeyFromEnv(keyName: string) {
+    const key = process.env[keyName];
+    if (!key) {
+        throw Error(`Missing ${keyName} in env`);
+    }
+
+    return key;
+}
